further optimization

+++ mcu/src/Main.c
@@ -75,7 +75,7 @@ void handle_message(circular_buf *buffer_ptr) {
     mess_len = (read_buffer(buffer_ptr) << 8);
     mess_len |= read_buffer(buffer_ptr);
 
-    for (uint16_t i = 0; i < mess_len; i++) {
+    for (size_t i = 0; i < mess_len; i++) {
         while(!( UCSR0A & (1<<UDRE0)));
         UDR0 = read_buffer(buffer_ptr);
     }
